Move internal elements to `internal` package.
